fix(ai): decouple tool result helpers from executor
