+++ src/lib/ai/tools/executor.ts
@@ -81,10 +81,20 @@ import type {
 } from "@/lib/ai/schedule-extraction";
 import { dispatchToolModule, getToolModule, isRegisteredTool } from "@/lib/ai/tools/registry";
 import {
+  toolError,
+  type ToolExecutionResult,
+} from "@/lib/ai/tools/result";
+import {
   buildMemberName,
   getSafeErrorMessage,
   safeToolQuery,
 } from "@/lib/ai/tools/shared";
+
+export type {
+  ScheduleFileToolErrorCode,
+  ToolExecutionErrorCode,
+  ToolExecutionResult,
+} from "@/lib/ai/tools/result";
 
 export type ToolExecutionAuthorization =
   | {
@@ -115,33 +125,6 @@ export interface ToolExecutionContext {
     mimeType: string;
   };
 }
-
-export type ScheduleFileToolErrorCode =
-  | "attachment_required"
-  | "invalid_attachment_path"
-  | "org_context_failed"
-  | "attachment_unavailable"
-  | "image_too_large"
-  | "image_timeout"
-  | "image_unreadable"
-  | "image_model_misconfigured"
-  | "pdf_unreadable"
-  | "pdf_timeout";
-
-export type ToolExecutionErrorCode =
-  | ScheduleFileToolErrorCode
-  | "enterprise_billing_role_required"
-  | "enterprise_invite_role_required"
-  | "pending_action_revise_limit"
-  | "pending_action_not_pending"
-  | "pending_action_conflict";
-
-export type ToolExecutionResult =
-  | { kind: "ok"; data: unknown }
-  | { kind: "forbidden"; error: "Forbidden" }
-  | { kind: "auth_error"; error: "Auth check failed" }
-  | { kind: "tool_error"; error: string; code?: ToolExecutionErrorCode }
-  | { kind: "timeout"; error: "Tool timed out" };
 
 type CountResult =
   | { ok: true; count: number }
@@ -417,10 +400,6 @@ function validateArgs(
     };
   }
   return { valid: true, args: parsed.data };
-}
-
-export function toolError(error: string, code?: ToolExecutionErrorCode): ToolExecutionResult {
-  return code ? { kind: "tool_error", error, code } : { kind: "tool_error", error };
 }
 
 function buildPendingActionField(
